I did a lot of things
* Added a settings file
* Added aliases, and a command to view them (!aliases)
* Bot class now stores its JDA, not its JDABuilder
* Moved much of the code in main to Bot.
* Added a method to Bot to get all visible users.
* Added methods to save/load settings.
* Overhauled the Command system
* Added !about
* Fixed a bug where RubixMaxim would crash when PM'd

+++ src/com/xetanai/rubix/MessageListener.java
@@ -1,6 +1,9 @@
 package com.xetanai.rubix;
 
 import net.dv8tion.jda.hooks.ListenerAdapter;
+
+import com.xetanai.rubix.Commands.Command;
+
 import net.dv8tion.jda.events.message.MessageReceivedEvent;
 
 public class MessageListener extends ListenerAdapter {
@@ -18,7 +21,8 @@ public class MessageListener extends ListenerAdapter {
         if (event.getMessage().getContent().startsWith(prefix))
         {
         	System.out.println("[MessageListener] Command received from "+ event.getAuthor().getUsername() +": "+ event.getMessage().getContent());
-        	event.getMessage().deleteMessage();
+        	if (bot.getSettings().deleteCommands() && !event.isPrivate())
+        		event.getMessage().deleteMessage();
         }
         else
         	return;
@@ -32,6 +36,16 @@ public class MessageListener extends ListenerAdapter {
         		return;
         	}
         }
+        for (Alias alias : bot.getAliasList())
+        {
+        	if (event.getMessage().getContent().startsWith(prefix+alias.getKeyword()))
+        	{
+        		System.out.println("[MessageListener] Command is an alias for "+ alias.getCommand().getKeyword() +". Executing that.");
+        		alias.getCommand().onCalled(bot, event);
+        		return;
+        	}
+        }
+        
         System.out.println("[MessageListener] Not a registered command. Ignoring.");
     }
 }
